+++ test/plugin/remove_invalid_test.rb
@@ -21,6 +21,12 @@ describe Shrine::Plugins::RemoveInvalid do
     assert_equal previous_attachment, @attacher.get
   end
 
+  it "removes the dirty state from the attacher" do
+    @attacher.class.validate { errors << :foo }
+    @attacher.set(cached_file = @attacher.cache!(fakeio))
+    refute @attacher.attached?
+  end
+
   it "doesn't remove stored files" do
     @attacher.class.validate { errors << :foo }
     @attacher.set(stored_file = @attacher.store!(fakeio))